:tractor: initial refactoring
