+++ pyecharts/render/engine.py
@@ -65,71 +65,55 @@ class RenderEngine:
         return utils.replace_placeholder(tpl.render(**kwargs))
 
 
-class Render:
-    def __int__(self):
-        self.js_host: str = ""
-        self.js_functions: utils.OrderedSet = utils.OrderedSet()
-        self.js_dependencies = utils.OrderedSet()
+def render(
+    chart, path: str, template_name: str, env: Optional[Environment], **kwargs
+) -> str:
+    RenderEngine(env).render_chart_to_file(
+        template_name=template_name, chart=chart, path=path, **kwargs
+    )
+    return os.path.abspath(path)
 
-    def _prepare_render(self):
-        raise NotImplementedError
 
-    def render_embed(self):
-        raise NotImplementedError
+def render_embed(
+    chart, template_name: str, env: Optional[Environment], **kwargs
+) -> str:
+    return RenderEngine(env).render_chart_to_template(
+        template_name=template_name, chart=chart, **kwargs
+    )
 
-    def _render(
-        self, path: str, template_name: str, env: Optional[Environment], **kwargs
-    ) -> str:
-        self._prepare_render()
-        RenderEngine(env).render_chart_to_file(
-            template_name=template_name, chart=self, path=path, **kwargs
+
+def render_notebook(self, notebook_template, lab_template):
+    instance = self if isinstance(self, Iterable) else (self,)
+    if CurrentConfig.NOTEBOOK_TYPE == NotebookType.JUPYTER_NOTEBOOK:
+        require_config = utils.produce_require_dict(
+            self.js_dependencies, self.js_host
         )
-        return os.path.abspath(path)
-
-    def _render_embed(
-        self, template_name: str, env: Optional[Environment], **kwargs
-    ) -> str:
-        self._prepare_render()
-        return RenderEngine(env).render_chart_to_template(
-            template_name=template_name, chart=self, **kwargs
+        return HTML(
+            RenderEngine().render_chart_to_notebook(
+                template_name=notebook_template,
+                charts=instance,
+                config_items=require_config["config_items"],
+                libraries=require_config["libraries"],
+            )
         )
 
-    def _render_notebook(self, notebook_template, lab_template):
-        instance = self if isinstance(self, Iterable) else (self,)
-        if CurrentConfig.NOTEBOOK_TYPE == NotebookType.JUPYTER_NOTEBOOK:
-            require_config = utils.produce_require_dict(
-                self.js_dependencies, self.js_host
+    if CurrentConfig.NOTEBOOK_TYPE == NotebookType.JUPYTER_LAB:
+        return HTML(
+            RenderEngine().render_chart_to_notebook(
+                template_name=lab_template, charts=instance
             )
-            return HTML(
-                RenderEngine().render_chart_to_notebook(
-                    template_name=notebook_template,
-                    charts=instance,
-                    config_items=require_config["config_items"],
-                    libraries=require_config["libraries"],
-                )
-            )
+        )
 
-        if CurrentConfig.NOTEBOOK_TYPE == NotebookType.JUPYTER_LAB:
-            return HTML(
-                RenderEngine().render_chart_to_notebook(
-                    template_name=lab_template, charts=instance
-                )
-            )
+    if CurrentConfig.NOTEBOOK_TYPE == NotebookType.NTERACT:
+        return HTML(self.render_embed())
 
-        if CurrentConfig.NOTEBOOK_TYPE == NotebookType.NTERACT:
-            return HTML(self.render_embed())
+    if CurrentConfig.NOTEBOOK_TYPE == NotebookType.ZEPPELIN:
+        print("%html " + self.render_embed())
 
-        if CurrentConfig.NOTEBOOK_TYPE == NotebookType.ZEPPELIN:
-            print("%html " + self.render_embed())
 
-    def load_javascript(self):
-        scripts = []
-        for dep in self.js_dependencies.items:
-            f, ext = FILENAMES[dep]
-            scripts.append("{}{}.{}".format(CurrentConfig.ONLINE_HOST, f, ext))
-        return Javascript(lib=scripts)
-
-    def add_js_funcs(self, *fns):
-        for fn in fns:
-            self.js_functions.add(fn)
-        return self
+def load_javascript(chart):
+    scripts = []
+    for dep in chart.js_dependencies.items:
+        f, ext = FILENAMES[dep]
+        scripts.append("{}{}.{}".format(CurrentConfig.ONLINE_HOST, f, ext))
+    return Javascript(lib=scripts)
