+++ requirements-dev.txt
@@ -1,2 +1,2 @@
-black==19.3b0
-isort==4.3.16
+black
+isort
